Add support for managing BodyType and MuscleBalance entities
Introduced services, controllers, and DTOs to handle CRUD operations for BodyType and MuscleBalance. Enhanced Minio integration by including bucket policy management and full path generation for files. Updated constants and listeners to improve error handling and file management.

+++ src/main/java/org/mohammed/cmsapi/listener/BodyTypeListener.java
@@ -3,8 +3,8 @@ package org.mohammed.cmsapi.listener;
 import jakarta.persistence.PrePersist;
 import jakarta.persistence.PreRemove;
 import jakarta.persistence.PreUpdate;
-import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.mohammed.cmsapi.config.properties.MinioProperties;
 import org.mohammed.cmsapi.constants.BodyTypeConstants;
 import org.mohammed.cmsapi.dto.MinioUploadedFileGetDto;
 import org.mohammed.cmsapi.exception.BodyTypeCreationException;
@@ -12,12 +12,19 @@ import org.mohammed.cmsapi.exception.BodyTypeDeleteException;
 import org.mohammed.cmsapi.exception.BodyTypeUpdateException;
 import org.mohammed.cmsapi.model.BodyType;
 import org.mohammed.cmsapi.service.MinioService;
+import org.springframework.stereotype.Component;
 
 @Slf4j
-@RequiredArgsConstructor
+@Component
 public class BodyTypeListener {
 
     private final MinioService minioService;
+    private final String prefixEndpoint;
+
+    public BodyTypeListener(MinioProperties minioProperties, MinioService minioService) {
+        this.minioService = minioService;
+        this.prefixEndpoint = String.format("%s/api/v1/buckets/%s/objects/download?preview=true&prefix=", minioProperties.getPreviewEndpoint(), BodyTypeConstants.BUCKET_NAME);
+    }
 
     @PrePersist
     public void uploadFileToMinio(BodyType bodyType) {
@@ -25,6 +32,7 @@ public class BodyTypeListener {
             if (bodyType.getId() != null) {
                 MinioUploadedFileGetDto uploadedFileGetDto = minioService.uploadFile(bodyType.getFile(), BodyTypeConstants.BUCKET_NAME);
                 bodyType.setImage(uploadedFileGetDto.object());
+                bodyType.setFullPath(prefixEndpoint + uploadedFileGetDto.object());
             }
         } catch (Exception e) {
             throw new BodyTypeCreationException("Error happened while Creating this Body Type");
@@ -34,10 +42,11 @@ public class BodyTypeListener {
     @PreUpdate
     public void updateFileInMinio(BodyType bodyType) {
         try {
-            if (bodyType.getId() != null) {
+            if (bodyType.getId() != null && bodyType.getFile() != null) {
                 MinioUploadedFileGetDto uploadedFileGetDto = minioService.uploadFile(bodyType.getFile(), BodyTypeConstants.BUCKET_NAME);
                 String oldImage = bodyType.getImage();
                 bodyType.setImage(uploadedFileGetDto.object());
+                bodyType.setFullPath(prefixEndpoint + uploadedFileGetDto.object());
                 minioService.deleteFile(BodyTypeConstants.BUCKET_NAME, oldImage);
             }
         } catch (Exception e) {
