+++ src/main/java/org/mohammed/cmsapi/listener/MuscleBalanceListener.java
@@ -3,8 +3,8 @@ package org.mohammed.cmsapi.listener;
 import jakarta.persistence.PrePersist;
 import jakarta.persistence.PreRemove;
 import jakarta.persistence.PreUpdate;
-import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.mohammed.cmsapi.config.properties.MinioProperties;
 import org.mohammed.cmsapi.constants.MuscleBalanceConstants;
 import org.mohammed.cmsapi.dto.MinioUploadedFileGetDto;
 import org.mohammed.cmsapi.exception.MuscleBalanceCreationException;
@@ -12,12 +12,19 @@ import org.mohammed.cmsapi.exception.MuscleBalanceDeleteException;
 import org.mohammed.cmsapi.exception.MuscleBalanceUpdateException;
 import org.mohammed.cmsapi.model.MuscleBalance;
 import org.mohammed.cmsapi.service.MinioService;
+import org.springframework.stereotype.Component;
 
 @Slf4j
-@RequiredArgsConstructor
+@Component
 public class MuscleBalanceListener {
 
     private final MinioService minioService;
+    private final String prefixEndpoint;
+
+    public MuscleBalanceListener(MinioProperties minioProperties, MinioService minioService) {
+        this.minioService = minioService;
+        this.prefixEndpoint = String.format("%s/api/v1/buckets/%s/objects/download?preview=true&prefix=", minioProperties.getPreviewEndpoint(), MuscleBalanceConstants.BUCKET_NAME);
+    }
 
     @PrePersist
     public void uploadFileToMinio(MuscleBalance muscleBalance) {
@@ -25,6 +32,7 @@ public class MuscleBalanceListener {
             if (muscleBalance.getId() != null) {
                 MinioUploadedFileGetDto uploadedFileGetDto = minioService.uploadFile(muscleBalance.getFile(), MuscleBalanceConstants.BUCKET_NAME);
                 muscleBalance.setImage(uploadedFileGetDto.object());
+                muscleBalance.setFullPath(prefixEndpoint + uploadedFileGetDto.object());
             }
         } catch (Exception e) {
             throw new MuscleBalanceCreationException("Error happened while Creating this Muscle Balance");
@@ -34,10 +42,11 @@ public class MuscleBalanceListener {
     @PreUpdate
     public void updateFileInMinio(MuscleBalance muscleBalance) {
         try {
-            if (muscleBalance.getId() != null) {
+            if (muscleBalance.getId() != null && muscleBalance.getFile() != null) {
                 MinioUploadedFileGetDto uploadedFileGetDto = minioService.uploadFile(muscleBalance.getFile(), MuscleBalanceConstants.BUCKET_NAME);
                 String oldImage = muscleBalance.getImage();
                 muscleBalance.setImage(uploadedFileGetDto.object());
+                muscleBalance.setFullPath(prefixEndpoint + uploadedFileGetDto.object());
                 minioService.deleteFile(MuscleBalanceConstants.BUCKET_NAME, oldImage);
             }
         } catch (Exception e) {
